**Add Docker installation script**
This commit adds a new shell script `get-docker.sh` that automates the installation of the latest stable versions of Docker CLI, Docker Engine, and their dependencies across various Linux distributions. The script:

- Detects the Linux distribution and version and configures the package management system accordingly
- Installs Docker CLI, Docker Engine, Docker Buildx, Docker Compose, containerd, and runc
- Provides options to install a specific version or from an alternative installation channel (e.g., "test")
- Supports installation from mirrors like Aliyun and AzureChinaCloud

This script is intended to simplify the Docker installation process and make it more accessible to users, especially those who are new to Docker or prefer a more automated approach.